test(#723): keep phase detail property coverage in existing suite
